Updated Dockerfile, Tweets, Posting, Nav
Added in DockerFile information for front and backend.
Added in basic navigation. Still have to add in proper icons, and connect the links to pages.
Added in basic tweet textfield with post button. Nothing is connected here yet.
Added in example tweet feed with one example tweet. Nothing is styled yet or connected to any backend.