Multiple Tweets Displayed, Main Page Formatting, Follow Button
Changed how tweets where displayed to allow for multiple tweets to appear as once, af it they were returned from the database.
Edited the main page to a grid system, and removed the drawer from the navigation bar, format is closer to standard twitter format.
Added in follow button to who to follow